Add pages management
The management allows the creation of pages using markdown as the body
of the page.
Also uses pretty urls for routing with automatic creation of slugs that can
also be edited.

Required fields of a page are:
 * title
 * description
 * body
 * author

The page body editor can make a preview of the content what allows better edition.